+++ radixdlt-regression/system-tests/src/test/java/com/radixdlt/test/chaos/ansible/AnsibleImageWrapper.java
@@ -2,7 +2,7 @@ package com.radixdlt.test.chaos.ansible;
 
 import com.radixdlt.test.RemoteBFTNetwork;
 import com.radixdlt.test.StaticClusterNetwork;
-import com.radixdlt.test.utils.ChaosExperimentUtils;
+import com.radixdlt.test.chaos.utils.ChaosExperimentUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.assertj.core.util.Lists;
